Add boilerplate for predictions module and tests.
This sets the basis of how the predictions module will be structured.
User will be able to call `model.predict()` in the future on an
arbitrary model. Most of this is boilerplate code, but it is important
for others to see now.

I have set up the tests to run with the `unittest`. This is built into
the standard python library. This allows us to run unit test through the
VSCode interface. Alternatively, you can still run them using `python -m
unittest`.

I have also set up the `gitignore` file to ignore the `__pycache__`.