Add definitions for CSFs.
These CSFs will be used to later identify values and correlations
between success metrics and the success of a project. We could go
without this, but for the sake of understanding what data is being
passed around, it'll be useful to have these definitions - especially
when inputs and outputs are being passed to and from the user.