Add success metrics.
Given critical success factors, we will predict the success metrics of a
project. I have labelled each with the question asked in the survey of
the research paper for ease of reference. I'll add the questions to the
CSFs in a separate PR. Those will be used to generate values for the
CSFs by the user to predict the success metrics on.