Add questions to CSF class.
These questions were used to gather the dataset that we are using, so it
is important to use these exact questions when we are trying to predict
success metrics. These should be shown to the managers when they are
trying to predict the success of a project at any given time.

Most of this is just adding text and no added functionality.