Create and train a KNN model.
This almost completes the KNN implementation. The only thing left is to
implement the predict method. This PR is a little large as it strings
together a lot of the pieces that were previously implemented.

The main changes are:
1. Fixing dataset splitting,
2. Handle missing values in the dataset.
3. Generic model trainer class,
4. KNN model training.

All of which piece together to allow us to train a KNN model. A seperate
model is used to predict each of the `SuccessMetrics`.