Complete the KNN model.
The KNN model is now complete and is ready to be used. As KNN training
is deterministic, no parameters are stored locally. However, if we wish
to do that to save about 3s upon starting the server, we can do that by
dumping `trained_params` into a json file and initializing the models
with the dataset.