Complete KNN model feedback.
We wish to give feedback about each success metrics as different
managers may have different priorities. We do this by mutating the
current CSFs of the project and then running the model again, in search
of a better SM score. The mutation is a simple +/- 1 to a single CSF.
This is to allow for simple feedback.