Generate data for report on predictions.
I've created a notebook that generates a report on the predictions using
the models that we've trained. Comparisons between random predictions,
predictions using the correlation based, baseline model (naive), and
the KNN model. The naive model can only output predictions for the
`overall_success` metric, whilst the other model can output predictions
for all 18 metrics.

The naive and KNN models performed significantly better than the random
predictions. The KNN model also beats the naive model for the
`overall_success` despite not directly predicting it - it predicts the
individual metrics that are used to calculate the `overall_success`.

This concludes that the KNN model may be a suitable model for the
predictions. However, it is worth noting that our dataset is very small,
with `n = 101`.