Change `model.give_feedback` to return JSON.
Previously, `model.give_feedback` returned a string, which gave feedback
for each SuccessMetric on a separate line. This isn't very useful for
the web interface, which needs to display the feedback in a table.

This commit changes `model.give_feedback` to return a JSON object, which
is easier to parse and display in the web interface.